Fix clippy
