+++ crates/dojo-indexer/src/tests/fixtures/entities.sql
